Created season_stat and season_stat_spec, removed methods and testing from stat tracker
